Refacto Structure Code
