+++ index.html
@@ -26,15 +26,16 @@
     <!--<p><a href="/visiteur/">Vous êtes un visiteur</a></p>-->
     
     <script src="http://code.jquery.com/jquery-1.10.1.min.js"></script>
-    <script src="/js/settings.js"></script>
-	<script>           
-            
+    <script src="/js/common_settings.js"></script>
+    <script src="/js/common_devSettings.js"></script>
+	<script>                    
             // On récupère depuis nodejs le nom de la machine serveur
             // en passant par de l'AJAX plutôt que par websocket....
             hostName = "???";    
             $.get( "/getvar", function( data ) {
               hostName = data.hostName;
-              if (hostName != "???" ) $('#zone_info_server').replaceWith(settings.appName() + " V " + settings.appVersion()+ " ("+hostName+")");
+              infoServerTxt = settings.appBranch() + ' - ' + settings.appName() + " V " + settings.appVersion()+ " ("+hostName+")";
+              if (hostName != "???" ) $('#zone_info_server').replaceWith(infoServerTxt);
             });
     </script>	
 </body>
