+++ js/common_settings.js
@@ -5,13 +5,15 @@ exports.appName = function(){
 };
 
 exports.appVersion = function(){
-    return '0.7.4.0';
-
+    return '0.7.4.1';
 };
 
 exports.appHostName = function(){
     return 'websocket-azkar';
 };
 
+exports.appBranch = function(){
+    return '1to1-thierry';
+};
 
 })(typeof exports === 'undefined'? this['settings']={}: exports);
