--- js/devSettings.js
@@ -1,8 +1,0 @@
-(function(exports){
-
-exports.appBranch = function(){
-    return '1to1-thierry';
-};
-
-
-})(typeof exports === 'undefined'? this['devSettings']={}: exports);
