+++ js/module_navigation.js
@@ -1,5 +1,9 @@
 // TODO ...
-// Ici on mettra tout ce qui concerne les canvas de télémétrie...
+// Ici on mettra tout ce qui concerne la navigation...
+// >> cartographie
+// >> liste des points d'intérets
+// >> fonctions de clic and go sur la cartographie
+// >> Affichage de la télémétrie
 
 var canvasPilote  = document.querySelector('#mapviewer_canvas');
 var ctxPilote = canvasPilote.getContext('2d');
@@ -7,10 +11,6 @@ var ctxPilote = canvasPilote.getContext('2d');
 
 var fondDemo = new Image();
 fondDemo.src = '/images/mapwiever1.png';
-console.log(fondDemo);
-//ctxPilote.drawImage(fondDemo, 620, 350, 0, 0);
-//alert("coucou");
-/**/
 
 
 fondDemo.addEventListener('load', function() {
@@ -19,10 +19,3 @@ fondDemo.addEventListener('load', function() {
 }, false);
 /**/
 
-/*
-ctxPilote.fillStyle = "gold";
-ctxPilote.fillRect(50, 35, 50, 80);
-
-ctxPilote.fillStyle = "rgba(23, 145, 167, 0.5)";
-ctxPilote.fillRect(40, 25, 40, 40);
-/**/
