+++ js/module_settings.js
@@ -1,0 +1,1 @@
+// TODO récupérer les paramétrages du benchmark
