+++ robot.html
@@ -161,18 +161,32 @@
         <script src="/socket.io/socket.io.js"></script>
         <script src='/js/lib/adapter.js'></script>
 		<script src="/js/lib/FileSaver.js"></script>
-        <script src="/js/common.js"></script>
-        <script src="/js/settings.js"></script>
-        <script src="/js/logs.js"></script>
-        <!--<script src='/js/battery.js'></script>-->
-        <script src='/js/robubox.js'></script>
-        <script src='/js/ws_manager.js'></script>
-		<script src='/js/main.js'></script>
+        
+        <script src="/js/common_tools.js"></script>
+        <script src="/js/common_settings.js"></script>
+        <script src='/js/common_robubox.js'></script>
 
-		<script>
+ 		<script src='/js/main_websocket.js'></script>
+ 		<script src='/js/main.js'></script>
+		
+		<script src="/js/module_logs.js"></script>
+
+
+
+		<script>                    
+            // On récupère depuis nodejs le nom de la machine serveur
+            // en passant par de l'AJAX plutôt que par websocket....
+            hostName = "???";    
+            $.get( "/getvar", function( data ) {
+              hostName = data.hostName;
+              //if (hostName != "???" ) $('#zone_info_server').replaceWith(settings.appName() + " V " + settings.appVersion()+ " ("+hostName+")");
+              infoServerTxt = settings.appBranch() + ' - ' + settings.appName() + " V " + settings.appVersion()+ " ("+hostName+")";
+              if (hostName != "???" ) $('#zone_info_server').replaceWith(infoServerTxt);
+            });
+    	
+			// On lance l'utilitaire de récupération de la batterie
 			robubox.getBattery();
 		</script>
-
 
 </body>
 </html>
